pyproject.toml
[tool.poetry]
name = "drf-project"
version = "0.1.0"
description = ""
authors = ["KenArsen <kenzhegulov.kgz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
djangorestframework = "^3.15.2"
pillow = "^11.1.0"
requests = "^2.32.3"
drf-spectacular = "^0.28.0"
drf-spectacular-sidecar = "^2025.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
